chore: fix some more pages
