+++ apps/admin/src/app/register/page.tsx
@@ -1,39 +1,9 @@
-import Head from "next/head";
-import React from "react";
-import { GoogleReCaptchaProvider } from "react-google-recaptcha-v3";
+import RegsiterWithCaptchaProvier from "./component";
 
-import { RegisterForm } from "@/components/register/views/RegisterForm";
+const Page = () => (
+  <RegsiterWithCaptchaProvier
+    recaptchaKey={process.env.NEXT_PUBLIC_RECAPTCHA_KEY}
+  />
+);
 
-const Register = () => {
-  return (
-    <>
-      <Head>
-        <title>Register</title>
-      </Head>
-      <div className="register" style={{ height: "100%", flex: 1 }}>
-        <RegisterForm />
-      </div>
-    </>
-  );
-};
-
-interface Props {
-  recaptchaKey: string;
-}
-const Provider: React.FC<Props> = ({ recaptchaKey }) => {
-  if (!recaptchaKey) return <Register />;
-  return (
-    <GoogleReCaptchaProvider reCaptchaKey={recaptchaKey}>
-      <Register />
-    </GoogleReCaptchaProvider>
-  );
-};
-export default Provider;
-
-export const getServerSideProps = async () => {
-  return {
-    props: {
-      recaptchaKey: process.env.RECAPTCHA_KEY_CLIENT || null,
-    },
-  };
-};
+export default Page;
