+++ apps/admin/src/components/navigation-builder/sortable/index.tsx
@@ -14,6 +14,7 @@ import {
   verticalListSortingStrategy,
 } from "@dnd-kit/sortable";
 import { FC, useState } from "react";
+import { BiPlus, BiSave } from "react-icons/bi";
 import { Button } from "ui";
 
 import { Collection } from "@/components/navigation-builder/types";
@@ -148,9 +149,21 @@ export const List: FC<Props> = ({ items = [], suggestions, onChange }) => {
           />
         ))}
       </SortableContext>
-      <div className="flex items-center justify-center">
-        <Button onClick={addNewRow} data-testid="newMenuBtn">
+      <div className="flex items-end justify-center gap-2">
+        <Button
+          onClick={addNewRow}
+          data-testid="newMenuBtn"
+          variant="dark"
+          className="flex items-center justify-center gap-1"
+        >
+          <BiPlus size={18} />
           Add New
+        </Button>
+        <Button
+          type="submit"
+          className="flex items-center justify-center gap-1"
+        >
+          <BiSave size={18} /> Save
         </Button>
       </div>
       <SuggestionModal
