+++ apps/admin/src/components/register/views/RegisterForm.tsx
@@ -1,5 +1,5 @@
 import Link from "next/link";
-import { useRouter } from "next/router";
+import { useRouter } from "next/navigation";
 import { signOut } from "next-auth/react";
 import React, { useEffect, useState } from "react";
 import { useGoogleReCaptcha } from "react-google-recaptcha-v3";
@@ -196,7 +196,7 @@ export const RegisterForm = () => {
                     href="/login"
                     className="text-blue-500 hover:underline focus:underline focus:outline-none"
                   >
-                    <a>Sign in</a>
+                    Sign in
                   </Link>
                   .
                 </p>
