+++ apps/admin/src/components/settings/general.tsx
@@ -1,32 +1,45 @@
 import { useFormContext } from "react-hook-form";
 import { Button, Input, TextArea } from "ui";
 
+import { Setting } from "../../../__generated__/__types__";
+
 interface Props {}
 const General: React.FC<Props> = () => {
-  const data = useFormContext();
+  const data = useFormContext<Setting>();
 
   return (
     <>
       <div className="mb-8 grid w-full grid-cols-1 gap-8 lg:grid-cols-2">
-        <Input
-          label="Site Title"
-          {...data.register("site_title", {
-            required: true,
-            maxLength: 30,
-          })}
-          placeholder="Name of your blog"
-          data-testid="siteTitle"
-          limit={30}
-        />
-
-        <Input
-          label="Site Tagline"
-          {...data.register("site_tagline", { required: true, maxLength: 60 })}
-          placeholder="A short phrase for your blog"
-          data-testid="siteTagline"
-          limit={60}
-        />
-
+        <div>
+          <Input
+            label="Site Title"
+            {...data.register("site_title", {
+              required: true,
+              maxLength: 30,
+            })}
+            placeholder="Name of your blog"
+            data-testid="siteTitle"
+            maxLength={30}
+          />
+          <div className="mt-2 text-right text-xs">
+            {data.watch("site_title")?.length}/30
+          </div>
+        </div>
+        <div>
+          <Input
+            label="Site Tagline"
+            {...data.register("site_tagline", {
+              required: true,
+              maxLength: 60,
+            })}
+            maxLength={60}
+            placeholder="A short phrase for your blog"
+            data-testid="siteTagline"
+          />
+          <div className="mt-2 text-right text-xs">
+            {data.watch("site_tagline")?.length}/60
+          </div>
+        </div>
         <Input
           label="Site Email (public)"
           {...data.register("site_email")}
@@ -39,19 +52,22 @@ const General: React.FC<Props> = () => {
           {...data.register("site_url", { required: true })}
           placeholder="URL of your site"
         />
-
-        <TextArea
-          label="Short Description"
-          {...data.register("site_description", {
-            required: true,
-            maxLength: 190,
-          })}
-          placeholder="Write something about your site. Will be used in SEO and other pages"
-          autoGrow={true}
-          data-testid="shortDescription"
-          limit={190}
-        />
-
+        <div>
+          <TextArea
+            label="Short Description"
+            {...data.register("site_description", {
+              required: true,
+              maxLength: 190,
+            })}
+            placeholder="Write something about your site. Will be used in SEO and other pages"
+            autoGrow={true}
+            data-testid="shortDescription"
+            maxLength={190}
+          />
+          <div className="mt-2 text-right text-xs">
+            {data.watch("site_description")?.length}/190
+          </div>
+        </div>
         <TextArea
           label="Footer Description"
           {...data.register("site_footer", { maxLength: 200 })}
