+++ apps/admin/src/pages/messages/registered.tsx
@@ -1,5 +1,5 @@
 import Head from "next/head";
-import { useRouter } from "next/router";
+import { useRouter } from "next/navigation";
 import { Button, Result } from "ui";
 
 const Registered = () => {
